refactor(calendar): split off week
